Update patch set 2
Patch Set 2: (6 inline comments)

Might I suggest adding the ability to make inline comments without JSON also? Perhaps as simple as the format that grep -n uses:

--inline 'file:line:message'

With an optional parameter, it might even be easy enough to supply a capturing regular expression to specify a specific inline format. I would think that patterns for many tools such as lint or other checkers would be easy to write small REs for. If not an RE, perhaps an even simpler printf style pattern where %f represents the filename, %l the line number, and %m the message. Example:

--inlineFormat '%f:[%l]%m'

Though I suspect that this would not be flexible enough without wildcard support. The java compiler would need something like

--inlineFormat '%f:\[%l,[0-9]\]%m'

since it prints the column after the linenumber. But maybe that is a job left to sed to prefilter?

Patch-set: 2