Update patch set 2
Patch Set 2:

Funny, I was working on the same thing, and it looks very similar to what Jyrki originally submitted in patchset#1 (new top level command), although mine is not that polished:
https://github.com/finik/gerrit/commit/393e734123f402dd8cfdd07ef4eaf9ea3fab2258

Is there any concern JSON and combining everything into one command might be an overkill for this?
For example, I currently have jenkins running:
git show HEAD | ./scripts/checkpatch.pl -
for every linux kernel commit. Output of kernel checkpatch.pl script is pretty simple (especially when --terse option is used), and I was planning to modify the script or to do a really simple post processing of the data and run a "gerrit comment" (or Jyrki's "gerrit inline") for each error (to create a draft comment) and then finish the session with existing "gerrit review" command to give it a final score and post all the drafts. Gathering all the failures and constructing a JSON for this is doable, but feels like an overkill.

Patch-set: 2
Reviewer: Dmitry Fink <1004137@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=0