Update patch set 10
Patch Set 10: Code-Review-1

I don't think this is ready yet.  During testing I've found that all the commit validations, including the plugin ones, get called twice on new patch sets.

This is not noticeable for the other validators because they raise exception if something is wrong.  But for the commit message length validation, messages are output.  This results in the message being output twice for new patch sets.

Patch-set: 10
Label: Code-Review=-1