Update patch set 6
Patch Set 6:

Thanks, this is pretty cool and obviously a lot of work. We appreciate the contribution.

A few of high-level issues before I dive into the code.

First of all, I can't get this to work on a local dev server. I click the button and the gutter widens, but nothing appears:

http://i.imgur.com/u6sX7jB.png

I am also concerned about performance. You did the first sensible thing, which is avoid doing work unless the user asks for the blame. But I am still concerned for several reasons:

First, potential for abuse. If your server is available on the public internet, anybody can DoS it just by making unauthenticated GET requests. This doesn't even require a malicious attacker--most search engine web crawlers these days will happily render JS apps and click all the links. The easiest way to ameliorate this is to have a configuration option to turn off blame for anonymous users.

Second, poor performance on large files. When we built Gitiles as a frontend for the Chromium repositories, we spent a lot of time implementing a cache for blame data. Otherwise the JGit (or even C git) blame implementation is just too slow to be usable. Try running `git blame` on this file:

https://chromium.googlesource.com/chromium/src/+blame/master/content/renderer/render_view_impl.cc

Even in the best case you're waiting several seconds for results to appear.

And during that time, the request is pegging a whole CPU. You run into extra trouble if the user starts clicking the blame icon repeatedly in frustration.

I think the performance issue is definitely surmountable by reusing the BlameCache from Gitiles, which is also based on JGit and has a compatible license. We might even be able to contribute it upstream directly into JGit (need to check the licensing situation though).

Nor do I think we need to block submission of this change until all the performance problems are worked out. As long as the functionality can be disabled as necessary.

Patch-set: 6