Update patch set 4
Patch Set 4: Code-Review-1

(4 comments)

High-level question: how are you planning on postfiltering users by visibility? In particular AccountControl#canSee might be pretty slow to call in a loop. It's not quite as crucial to do this super quickly as it was for changes, because there's no accounts equivalent of the search results/dashboard, but it might be worth thinking about.

Patch-set: 4
Label: Code-Review=-1