Update patch set 2
Patch Set 2:

(1 comment)

I think using another query parameter to switch between suggest and normal queries is fine. I wonder if it would make more sense as an option like ListChangesOption. On the one hand that avoids yet another parameter (if/when we add additional options in the future). On the other hand there is no equivalent ListChangesOption that actually changes the interpretation of the query, so maybe it's not quite applicable. What do you think?

Patch-set: 2
Label: Verified=0