Update patch set 2
Patch Set 2:

(1 comment)

> I think using another query parameter to switch between suggest and
 > normal queries is fine. I wonder if it would make more sense as an
 > option like ListChangesOption. On the one hand that avoids yet
 > another parameter (if/when we add additional options in the
 > future). On the other hand there is no equivalent ListChangesOption
 > that actually changes the interpretation of the query, so maybe
 > it's not quite applicable. What do you think?

Good point to consider, but I think I prefer the query parameter. ListChangesOption are "Output options per change", that control which information is included in the result. The suggest option is rather an input option to control how the query should be executed, and thus it seems to be more similar to the start and limit options, which are given as query parameters as well. Still we may want to add ListAccountsOption later, e.g. to control if detailed account information should be returned or not.

Patch-set: 2
Label: Verified=0