Add support for mariadb as a db in gerrit
This fixes support for allowing users to set character encoding and
collation such as utf8mb4 and utf8mb4_unicode_ci which currently with
mysql's you have to set the db to do it. But in an enterprise world you
can not do that or it is hard to do that if you share other db's on the
same host that use different character encoding and collation.

Please see our downstream report https://phabricator.wikimedia.org/T145885

We need full unicode to support emoji's as utf8 is not full unicode.

+ i found this https://mathiasbynens.be/notes/mysql-utf8mb4 which
says that this could be leaking data and be a security problem with utf8.

So i have no idea what security problems it holds for gerrit.

This plugin is also updated more regulary then mysql so more bug fixes
+ features.

See https://github.com/MariaDB/mariadb-connector-j/

https://downloads.mariadb.com/Connectors/java/connector-java-1.5.7/sha1sums.txt

Change-Id: I7cb80539cb8480b306e6fac1511a0bef34b5d4dd
Note: This will most likly remove support for mysql 5.4 or lower as we are replacing the mysql-connector plugin with the mariadb one.