SYSTEM (/) Support disabling change ReviewDb table access entirely
We already had DisabledChangesReviewDbWrapper to throw an exception
indicating it's a programmer error to try to read from changes tables
directly. By design, it was still possible to punch through this wrapper
and get access to the ReviewDb table.

We need to go deeper if we really want to shut off the changes tables.
Add a new config option that adds another level of wrapping to ReviewDb,
which makes all operations into complete no-ops. When this option is
set, there is literally no way to write to or read from the underlying
database table using the ReviewDb API. The table could go away and
Gerrit would still function.

Of course, extreme caution should be used before turning on this option.
At a minimum it is up to the admin to ensure that all changes in the
database have NoteDb as their primary storage location.

Change-Id: dns.code.develper@gmail.com  

Jordan Kneisley