Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Did something get broken by the ones that I submitted earlier?
> 
> We have an internal plugin that failed to compile because it used an API on RefControl that I hid with the series. Because I didn't give Dave an internal change that atomically updated both that hiding and the plugin at the same time, he's a bit frustrated.

I don't personally mind fixing stuff in general, but tonight I am frustrated because I'm trying to fix production and you're making it harder.

> I've been trying to atomically update core plugins when I hide APIs, but I haven't cared about open source non-core plugins. Dave is saying I now have to care about atomically merging our internal, proprietary, non-open source, non-core plugins.
> 
> A better strategy may be for me to try and fix all of our internal usages of *Control classes to PermissionBackend first, and then dispense with this silly notion of using Code-Review-2 to prevent submission.

That may work for this case, but this is a general problem we have anytime we change a public API or other functionality in a way that needs special care for googlesource.com. This API is on the simple end of the spectrum, but take a look at the internal CL that Edwin submitted earlier today that required hundreds of lines of internal changes. It's annoying for him to have to avoid submitting upstream changes until he's done ironing out those hundreds of lines worth of kinks, which takes a few days, yes. But it's far worse for velocity of imports for him to submit first and then it's impossible for anybody to import upstream Gerrit at all for those two days.

So now we have an internal policy that says: if you break an upstream API in a way that will require changes to googlesource.com, you the author are the most likely person to know about it, and the most likely person to know how to fix it, so it's on you to -2.

Patch-set: 3
Reviewer: Dave Borowitz <1010008@173816e5-2b9a-37c3-8a2e-48639d4f1153>