Display whole commit width up to 72ch if screen size allows
Previously, when the screen size shrank, the commit message and related
changes both got smaller. This caused the commit message to be difficult
to read.

This change sets flex-shrink to 0 on the commit container so that the
related changes absorb all of the shrinkage instead. Because this can,
in certain cases, cause related changes to be unreadable, two steps have
been taken to alleviate this:
  1) On hover of related change, display full text
  2) Add a new breakpoint at 60em, which moves related change below the
     commit message, before entering the full mobile view.

The rule to determine if the related changes toggle is needed is updated
to take into account this new view, so that it fits without jank
assuming there is space, and if not adds a minimal height that can be
expanded.

Bug: Issue 5706
Change-Id: I1ed723af55e9ca1706c8ef877902ee2c2e4148b2