Update patch set 2
Patch Set 2:

> Nice, thank you! I wonder how that pathological approach of reading of thousands of branches, where only one is needed, was undiscovered for years?

For most servers, this behavior even at 1000 branches wasn't actually that bad.

We only noticed a problem because internally this uses getRefs("refs/heads/") which is not efficient at googlesource.com. Its internally implemented as getRefs(""), which also has to dredge through refs/changes/ space. A repository had a lot of changes, and was starting to be accessed by this API from a plugin's UiAction.getDescription() method.
 
> This should be definitely done at least on stable-2.14 if not on stable-2.13.

There are quite a few merge conflicts for 2.14 and 2.13. Given that its gone so unnoticed for so long, I'm not sure if its worth the effort to resolve those 2 or 3 times (rebasing back to 2.13, merge to 2.14, merge to master).

Patch-set: 2
Reviewer: Shawn Pearce <1001240@173816e5-2b9a-37c3-8a2e-48639d4f1153>