Add --only-split flag to crisper call
Building PG concatenates via Vulcanize and Crisper. Currently the
resulting HTML file includes a script tag for gr-app.js (at the end).

To be able to apply a nonce to the script tag, the built HTML file
should not include this <script> tag so that gr-app.js can instead be
loaded via a script tag in the index itself. The --only-split flag
acccomplishes this task.

Bug: Issue 6684
Change-Id: I77d6617de655f3100b8cd615816b63d841471b3a