Show full date when message is expanded
The full date option is set by passing a show-full-date boolean to
gr-date-formatter. It should be noted that changed this value after the
date has rendered does not force a re-render. If it was initialized as
false and used as a parameter in the computed property, the date
flickers. If it is not initialized, then an undefined value is not
allowed. Since most formatted dates do not show this long form, adding
show-long-form="false" doesn't make sense.

Change-Id: I4524918a977e7fd4a557bf29fa834e67ceed9341